==> 1. Data Collection (News).ipynb <==
import requests
import datetime

def save_page(url, file_name):

    # Send a GET request to the URL and retrieve the response
    response = requests.get(url)

    # Check if the request was successful
    if response.status_code == 200:
        # Save the response content to a file
        with open(f"{file_name}.html", "wb") as f:
            f.write(response.content)
        print("Page downloaded successfully.")
    else:
        print(f"Error downloading page. Status code: {response.status_code}")


# url = "https://www.understandingwar.org/backgrounder/russian-offensive-campaign-assessment-may-7"
# save_page(url, "Reports/test2") # Create 'Reports' folder before

# Set the start date to February 24, 2022
start_date = datetime.date(2022, 2, 24)

# Get the current date
end_date = datetime.date.today()

# Define a list of month names
month_names = ['january', 'february', 'march', 'april', 'may', 'june', 'july', 'august', 'september', 'october', 'november', 'december']

# Iterate over the dates from the start date to the end date
for date in (start_date + datetime.timedelta(n) for n in range((end_date - start_date).days)):
    # Get the month name for the current date
    month_name = month_names[date.month - 1]

    day = int(date.strftime("%d"))
    name = date.strftime("%Y-%m-%d")
    # Print the date in the format "Month Day, Year" str(int(s))
    if date.strftime("%Y")== "2022":
        if date == datetime.date(2022, 5, 5):
            url = "https://www.understandingwar.org/backgrounder/russian-campaign-assessment-may-5"
        elif date == datetime.date(2022, 7, 11):
            url = "https://www.understandingwar.org/backgrounder/russian-offensive-campaign-update-july-11"
        elif date == datetime.date(2022, 8, 12):
            url = "https://www.understandingwar.org/backgrounder/russian-offensive-campaign-assessment-august-12-0"
        elif date > datetime.date(2022, 2, 28):
            url = ("https://www.understandingwar.org/backgrounder/russian-offensive-campaign-assessment-%s-%d"%(month_name, day))
        elif date == datetime.date(2022, 2, 24):
            url = "https://www.understandingwar.org/backgrounder/russia-ukraine-warning-update-initial-russian-offensive-campaign-assessment"
        elif date == datetime.date(2022, 2, 25):
            url = "https://www.understandingwar.org/backgrounder/russia-ukraine-warning-update-russian-offensive-campaign-assessment-february-25-2022"
        elif date == datetime.date(2022, 2, 28):
            url = "https://www.understandingwar.org/backgrounder/russian-offensive-campaign-assessment-february-28-2022"
        else:
            url = ("https://www.understandingwar.org/backgrounder/russia-ukraine-warning-update-russian-offensive-campaign-assessment-%s-%d"%(month_name, day))
    else:
        if date == datetime.date(2023, 2, 5):
            url = "https://www.understandingwar.org/backgrounder/russian-offensive-campaign-update-february-5-2023"
        else:
            year = date.strftime("%Y")
            url = ("https://www.understandingwar.org/backgrounder/russian-offensive-campaign-assessment-%s-%d-%s"%(month_name, day, year))
    print(url)
    save_page(url, f"Reports/{name}")
