==> 2.3. Data Preprocessing.ipynb <==
import datetime
import calendar
import numpy as np
import pandas as pd

import pickle

from sklearn.feature_extraction.text import TfidfTransformer

pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)

REPORTS_DATA_FOLDER = "data/2_isw_preprocessed"
REPORTS_DATA_FILE = "all_days.csv"

TFIDF_NUMBER = 100

EVENTS_DATA_FOLDER = "data/1_events"
EVENTS_DATA_FILE = "alarms.csv"

WEATHER_DATA_FOLDER = "data/1_weather"
WEATHER_DATA_FILE = "all_weather_by_hour_v2.csv"

REGIONS_DATA_FOLDER = "data/0_meta"
REGIONS_DATA_FILE = "regions.csv"

MODEL_FOLDER = "model"

OUTPUT_FOLDER = "data/4_all_data_preprocessed"
ISW_OUTPUT_DATA_FILE = "all_isw.csv"
WEATHER_EVENTS_OUTPUT_DATA_FILE = "all_hourly_weather_events_v2.csv"
WEATHER_EVENTS_KEYWORDS_OUTPUT_DATA_FILE = "all_hourly_weather_events_isw_v2.csv"

tfidf_transformer_model = "tfidf_transformer"
count_vectorizer_model = "count_vectorizer"

tfidf_transformer_version = "v4"
count_vectorizer_version = "v4"

def isNaN(num):
    return num != num

## prepare events data

# read events data
df_events = pd.read_csv(f"{EVENTS_DATA_FOLDER}/{EVENTS_DATA_FILE}", sep=";")

# evets dataset uses differet ids for the regions
df_events_v2 = df_events.drop(["id","region_id"],axis=1)
df_events_v2.sample(5)

# calculate descrete hours for event start and end
df_events_v2["start_time"] = pd.to_datetime(df_events_v2["start"])
df_events_v2["end_time"] = pd.to_datetime(df_events_v2["end"])

df_events_v2["start_hour"] = df_events_v2['start_time'].dt.floor('H')
df_events_v2["end_hour"] = df_events_v2['end_time'].dt.ceil('H')

df_events_v2["start_hour"] = df_events_v2.apply(lambda x: x["start_hour"] if not isNaN(x["start_hour"]) else x["event_hour"] , axis=1)
df_events_v2["end_hour"] = df_events_v2.apply(lambda x: x["end_hour"] if not isNaN(x["end_hour"]) else x["event_hour"] , axis=1)

# Present date as separate colum
df_events_v2["day_date"] = df_events_v2["start_time"].dt.date

# Preset tiime as epoch time (faster ops)
df_events_v2["start_hour_datetimeEpoch"] = df_events_v2['start_hour'].apply(lambda x: int(calendar.timegm(x.timetuple()))  if not isNaN(x) else None)
df_events_v2["end_hour_datetimeEpoch"] = df_events_v2['end_hour'].apply(lambda x: int(calendar.timegm(x.timetuple()))  if not isNaN(x) else None)

df_events_v2.head(5)

## prepare weather

# read weather time series set
df_weather = pd.read_csv(f"{WEATHER_DATA_FOLDER}/{WEATHER_DATA_FILE}")
df_weather["day_datetime"] = pd.to_datetime(df_weather["day_datetime"])

print(df_weather.shape)
df_weather.head(5)

# Reducing number of features and duplicate data
weather_exclude = [
    "day_feelslikemax",
    "day_feelslikemin",
    "day_sunriseEpoch",
    "day_sunsetEpoch",
    "day_description",
    "city_latitude",
    "city_longitude",
    "city_address",
    "city_timezone",
    "city_tzoffset",
    "day_feelslike",
    "day_precipprob",
    "day_snow",
    "day_snowdepth",
    "day_windgust",
    "day_windspeed",
    "day_winddir",
    "day_pressure",
    "day_cloudcover",
    "day_visibility",
    "day_severerisk",
    "day_conditions",
    "day_icon",
    "day_source",
    "day_preciptype",
    "day_stations",
    "hour_icon",
    "hour_source",
    "hour_stations",
    "hour_feelslike"
]

df_weather_v2 = df_weather.drop(weather_exclude, axis=1)

# differet formats for different cities :/
df_weather_v2["city"] = df_weather_v2["city_resolvedAddress"].apply(lambda x: x.split(",")[0])
df_weather_v2["city"] = df_weather_v2["city"].replace('Хмельницька область', "Хмельницький")

print(df_weather_v2.shape)
df_weather_v2.head(5)

## merging data

# Use region ids from regions dataset
df_regions = pd.read_csv(f"{REGIONS_DATA_FOLDER}/{REGIONS_DATA_FILE}")
df_regions.head(5)

df_weather_reg = pd.merge(df_weather_v2, df_regions, left_on="city",right_on="center_city_ua")

print(df_weather_reg.shape)
df_weather_reg.head(5)

### Merging weather and events

print(df_events_v2.shape)
df_events_v2.head(5)

# present evets as dictionary
events_dict = df_events_v2.to_dict('records')
events_by_hour = []

events_dict[0]

# Slice events into descrete samples one hour long each
for event in events_dict:
    for d in pd.date_range(start=event["start_hour"], end=event["end_hour"], freq='1H'):
        et = event.copy()
        et["hour_level_event_time"] = d
        events_by_hour.append(et)

df_events_v3 = pd.DataFrame.from_dict(events_by_hour)
df_events_v3["hour_level_event_datetimeEpoch"] = df_events_v3["hour_level_event_time"].apply(lambda x: int(calendar.timegm(x.timetuple()))  if not isNaN(x) else None)

print(df_events_v3.shape)
df_events_v3.head(5)

# merge events to weather data
df_events_v4 = df_events_v3.copy().add_prefix('event_')

df_weather_v4 = df_weather_reg.merge(df_events_v4, 
                                     how="left", 
                                     left_on=["region_alt","hour_datetimeEpoch"],
                                     right_on=["event_region_title","event_hour_level_event_datetimeEpoch"])

print(df_weather_v4.shape)
df_weather_v4.sample(5)

# Save weather data
df_weather_v4.to_csv(f"{OUTPUT_FOLDER}/{WEATHER_EVENTS_OUTPUT_DATA_FILE}", sep=";", index=False)

### Calculate top keywords for isw keywords (deprecated for model training; EDA use only)

# read reports dataset
df_isw = pd.read_csv(f"{REPORTS_DATA_FOLDER}/{REPORTS_DATA_FILE}", sep=";")
df_isw.head(5)

# load tfidf transformer
tfidf = pickle.load(open(f"{MODEL_FOLDER}/{tfidf_transformer_model}_{tfidf_transformer_version}.pkl", "rb"))
cv = pickle.load(open(f"{MODEL_FOLDER}/{count_vectorizer_model}_{count_vectorizer_version}.pkl", "rb"))


# Converts doument lemm to keywords vector, returs TFIDF_NUMBER top keywords as dict
def convert_to_vector(doc, tfidf_transformer, count_vectorized):
    feature_names = count_vectorized.get_feature_names_out()
    tfidf_vector = tfidf_transformer.transform(count_vectorized.transform([doc]))

    # sort vector by tf-idf score of each keyword
    keywords = pd.DataFrame(np.transpose(tfidf_vector.todense()), index=feature_names, columns=['tfidf'])
    keywords = keywords['tfidf'].sort_values(ascending=False).head(TFIDF_NUMBER).round(3).to_dict()
    
    return keywords

# Add top keywords to dataset
df_isw['keywords'] = df_isw['lemm'].apply(lambda x: convert_to_vector(x, tfidf, cv))
df_isw.head(5)

# Create tomorrow date column for merging
df_isw["date_datetime"] = pd.to_datetime(df_isw["date"])
df_isw['date_tomorrow_datetime'] = df_isw['date_datetime'].apply(lambda x: x+datetime.timedelta(days=1))
df_isw = df_isw.rename(columns = {"date_datetime":"report_date"})
df_isw.sample(5)

# Save dataset
df_isw.to_csv(f"{OUTPUT_FOLDER}/{ISW_OUTPUT_DATA_FILE}", sep=";", index=False)

# df_weather_v5 = df_weather_v4.merge(df_isw[["date_tomorrow_datetime", "keywords"]].add_prefix('isw_'),
#                                     how="left",
#                                     left_on="day_datetime",
#                                     right_on="isw_date_tomorrow_datetime")
# df_weather_v5.head(10)

# df_weather_v5.head(5)

# df_weather_v5.to_csv(f"{OUTPUT_FOLDER}/{WEATHER_EVENTS_KEYWORDS_OUTPUT_DATA_FILE}", sep=";", index=False)
