==> 5. Model Training.ipynb <==
import pandas as pd
import numpy as np
import statsmodels.api as sm

from sklearn.model_selection import TimeSeriesSplit

df_weather_events = pd.read_csv(f"data/4_all_data_preprocessed/all_hourly_weather_events_v2.csv", sep=";", parse_dates=['day_datetime'])
df_isw = pd.read_csv(f"data/4_all_data_preprocessed/all_isw.csv", sep=";")

# Sort data by date
df_weather_events = df_weather_events.sort_values(by='day_datetime')

df_weather_events.head()

df_weather_events.info()

tss = TimeSeriesSplit(n_splits=2)

train_splits, test_splits = tss.split(df_weather_events)
train_split = train_splits[1]
test_split = test_splits[1]

print("Train Split:", train_split)
print("Test Split:", test_split)

train_df = df_weather_events.iloc[train_split, :]
test_df = df_weather_events.iloc[test_split, :]

print("Train:", train_df.city_resolvedAddress.unique())
print("Test:", test_df.city_resolvedAddress.unique())
print("Train:", train_df.day_datetime.min(), train_df.day_datetime.max())
print("Test:", test_df.day_datetime.min(), test_df.day_datetime.max())

min_date = df_weather_events.day_datetime.min()
max_date = df_weather_events.day_datetime.max()
print("Min:", min_date, "Max:", max_date)

train_percent = .75
time_between = max_date - min_date
train_cutoff = min_date + train_percent*time_between
train_cutoff

train_df = df_weather_events[df_weather_events.day_datetime <= train_cutoff]
test_df = df_weather_events[df_weather_events.day_datetime > train_cutoff]

print("Train:", train_df.city_resolvedAddress.unique())
print("Test:", test_df.city_resolvedAddress.unique())
print("Train:", train_df.day_datetime.min(), train_df.day_datetime.max())
print("Test:", test_df.day_datetime.min(), test_df.day_datetime.max())

days_between = time_between / np.timedelta64(1, 'D')
days_between

train_days = 273
train_cutoff = min_date + pd.DateOffset(train_days)

train_df = df_weather_events[df_weather_events.day_datetime <= train_cutoff]
test_df = df_weather_events[df_weather_events.day_datetime > train_cutoff]

print("Train:", train_df.city_resolvedAddress.unique())
print("Test:", test_df.city_resolvedAddress.unique())
print("Train:", train_df.day_datetime.min(), train_df.day_datetime.max())
print("Test:", test_df.day_datetime.min(), test_df.day_datetime.max())
